Option to select timezone when creating battles
As the timezone information is not stored in replays, the only
way is to ask (or guess) the uploader for the correct timezone.
When creating a battle, the uploader now has to select the timezone,
which will be remembered in his session cookie for the future.
The date from the replay is then assumed to be in the selected time
zone and will be converted to the configurable server timezone.

Adjusting old battles in the database has to be done manually.